requirements.txt
openai-whisper==20231117
pyaudio==0.2.11
numpy==1.24.3
torch>=2.0.0

